[neutron] Add rabbitmq to neutron-chart
In order to migrate away from the openstack-helm rabbitmq,
first create a rabbitmq to be used.
As a next step, the service entry of the central rabbitmq will point
to this rabbit.
Then we can configure everything to use the new rabbitmq directly.
Finally, we can tear down the central rabbitmq.